+++ src/main/java/connect4/Main.java
@@ -1,10 +1,17 @@
 package connect4;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.Scanner;
+import java.util.stream.Collectors;
 
+import connect4.loader.BoardLoader;
 import connect4.player.ConsoleHumanPlayer;
 import connect4.player.Player;
 import connect4.player.TrainedComputerPlayer;
+import connect4.trainer.BoardAnalysis;
+import connect4.trainer.ColumnAnalysis;
+import connect4.trainer.Trainer;
 
 /**
  * Start that sucker!
@@ -30,7 +37,7 @@ public class Main {
 		if ("1".equals(choice)) {
 			playGame();
 		} else if ("2".equals(choice)) {
-			anaylseBoard();
+			anaylseBoard(scanner);
 		}
 		scanner.close();
 	}
@@ -47,8 +54,41 @@ public class Main {
 		controller.startGame();
 	}
 
-	private static void anaylseBoard() {
-		// TODO Auto-generated method stub
+	private static void anaylseBoard(final Scanner scanner) {
+		File boardFile = null;
+		while (boardFile == null || !boardFile.canRead()) {
+			System.out.print("Enter path to file of a board to analyse: ");
+			boardFile = new File(scanner.next().trim());
+			if (!boardFile.canRead()) {
+				System.out.println(String.format("Hmmm, couldn't load file '%s'. Enter a new file: ", boardFile.getAbsolutePath()));
+			}
+		}
+		System.out.println(String.format("Analysing board '%s'...", boardFile.getAbsolutePath()));
+		Board board;
+		try {
+			board = BoardLoader.readBoard(boardFile);
+		} catch (final IOException e) {
+			System.out.println(
+					String.format("Can't load board from file '%s' because: %s", boardFile.getAbsolutePath(), e.getLocalizedMessage()));
+			return;
+		}
+		System.out.println(board.toString(true));
+		final Trainer trainer = new Trainer();
 
+		for (final Disc disc : Disc.values()) {
+			System.out.println(String.format("Analysing board for %s...", disc.toString()));
+			final int column = trainer.recommend(board, disc);
+			final BoardAnalysis lastBestBoardAnalysis = trainer.getLastBestBoardAnalysis();
+			if (lastBestBoardAnalysis.size() == 1) {
+				System.out.println(String.format("Recommended move is column %d (column[%d])", column + 1, column));
+			} else {
+				System.out.println(String.format("Recommended columns are: %s", lastBestBoardAnalysis.stream()
+						.map(n -> (n.getColumn() + 1) + " (column[" + n.getColumn() + "])").collect(Collectors.joining(", "))));
+			}
+			for (final ColumnAnalysis columnAnalysis : lastBestBoardAnalysis) {
+				System.out.println(columnAnalysis.toStringDetail());
+			}
+			System.out.println();
+		}
 	}
 }